requirements.txt
streamlit==1.44.1
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.6.1
scipy==1.15.2
spacy==3.8.4
sentence-transformers==3.4.1
pdfplumber==0.11.5
python-docx==1.1.2
PyPDF2==3.0.1
rapidfuzz==3.12.1
plotly==6.0.1
pytest==8.3.5
